UPDATE: Change u32 to usize to work better with constants

+++ src/main.rs
@@ -2,7 +2,7 @@ use std::ops::Range;
 use std::process;
 use rayon::prelude::*;
 
-const fn numbers_are_unique(numbers: &[u32; 9]) -> bool {
+const fn numbers_are_unique(numbers: &[usize; 9]) -> bool {
     !(numbers[0] == numbers[1]
         || numbers[0] == numbers[2]
         || numbers[0] == numbers[3]
@@ -41,32 +41,32 @@ const fn numbers_are_unique(numbers: &[u32; 9]) -> bool {
         || numbers[7] == numbers[8])
 }
 
-const fn sums_are_equal(numbers: &[u32; 9]) -> bool {
-    let top_row_sum: u32 = numbers[0] + numbers[1] + numbers[2];
-    let middle_row_sum: u32 = numbers[3] + numbers[4] + numbers[5];
-    let bottom_row_sum: u32 = numbers[6] + numbers[7] + numbers[8];
+const fn sums_are_equal(numbers: &[usize; 9]) -> bool {
+    let top_row_sum: usize = numbers[0] + numbers[1] + numbers[2];
+    let middle_row_sum: usize = numbers[3] + numbers[4] + numbers[5];
+    let bottom_row_sum: usize = numbers[6] + numbers[7] + numbers[8];
 
     if top_row_sum != middle_row_sum || middle_row_sum != bottom_row_sum {
         return false;
     }
 
-    let left_column_sum: u32 = numbers[0] + numbers[3] + numbers[6];
-    let middle_column_sum: u32 = numbers[1] + numbers[4] + numbers[7];
-    let right_column_sum: u32 = numbers[2] + numbers[5] + numbers[8];
+    let left_column_sum: usize = numbers[0] + numbers[3] + numbers[6];
+    let middle_column_sum: usize = numbers[1] + numbers[4] + numbers[7];
+    let right_column_sum: usize = numbers[2] + numbers[5] + numbers[8];
 
     if bottom_row_sum != left_column_sum || left_column_sum != middle_column_sum || middle_column_sum != right_column_sum {
         return false;
     }
 
-    let nw_se_sum: u32 = numbers[0] + numbers[4] + numbers[8];
-    let sw_ne_sum: u32 = numbers[6] + numbers[4] + numbers[2];
+    let nw_se_sum: usize = numbers[0] + numbers[4] + numbers[8];
+    let sw_ne_sum: usize = numbers[6] + numbers[4] + numbers[2];
 
     right_column_sum == nw_se_sum && nw_se_sum == sw_ne_sum
 }
 
 fn main() {
-    const LIMIT: u32 = 60;
-    const NUMBER_ITERATOR: Range<u32> = 0..LIMIT;
+    const LIMIT: usize = 60;
+    const NUMBER_ITERATOR: Range<usize> = 0..LIMIT;
 
     NUMBER_ITERATOR.for_each(|a| {
         NUMBER_ITERATOR.into_par_iter().for_each(|b| {
